Use stdx expand
