+++ helix-loader/Cargo.toml
@@ -31,7 +31,6 @@ log = "0.4"
 cc = { version = "1" }
 threadpool = { version = "1.0" }
 tempfile.workspace = true
-regex = "1.11.1"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 libloading = "0.8"
